Move rapier systems into rapier_integration
No errors with clippy!
